Tweak setting of target temperatures
When target state is set to AUTO, we are assumed to be in home mode.  In
this case cooling and heating thresholds are set for the home activity.
Since both cooling and heating `set` listeners are triggered (in that
order), we cache the cooling temperature and set both when the heating
listener is called.

When target state is not AUTO, we are assumed to be in manual mode. In
this case we set both cooling and heating set points to be the same
temperature for the manual activity.