Update InfinitudeThermostat.js
Adding FilterLifeLevel to display 'filtrlvl'. This creates the possibility of later adding FilterChangeIndication for notifications on filter life. 

Note that 'filtrlvl' runs from 0-100 where the integers are % of life used, i.e., 0% means filter is new, 100% means Filter needs to be changed. I believe this is time-based and once it reaches 100% the expected behavior is for a user to change the filter and reset the alert on the thermostat, resetting to 0.